Explicitly Increase Guava Dependency Version and Adopt Stopwatch Usage
The Timing trait made use of deprecated Stopwatch calls based on the old Guava version being transitively loaded with the gdata library.
Relying on these deprecated methods makes it impossible to use the library in more recent project, if these include libraries that rely on a Guava version where the deprecated methods are already removed. This results in a java.lang.IllegalAccessError at runtime because the old Guava library the Timer trait relies on might be evicted by a potential more recent version.
This commit explicitly includes the latest Guava version and adopts the usage of the Stopwatch class in the Timing trait accordingly.